% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-rank.R
\name{data.sim_4_rank}
\alias{data.sim_4_rank}
\title{data.sim_4_rank}
\usage{
data.sim_4_rank(item.par, theta, BID)
}
\arguments{
\item{item.par}{a data.frame including a,d parameter}

\item{theta}{matrix N x D}

\item{BID}{block,item,dimension table}
}
\value{
N*J matrix.
}
\description{
data.sim_4_rank
}
