% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-mole.R
\name{kernel_4_mole}
\alias{kernel_4_mole}
\title{kernel_4_mole}
\usage{
kernel_4_mole(
  Y,
  BID,
  a,
  d,
  sigm,
  theta,
  B = 20,
  J,
  D,
  N,
  cor.matrix = FALSE,
  positive,
  fix.sigma = FALSE
)
}
\arguments{
\item{Y}{a # of subjects x # of blocks matrix of item responses}

\item{BID}{a # of statements x 3 matrix of item information; Columns are "Block", "Item" and "Dimensions"}

\item{a}{initial alpha parameters ; a vector with length = # of statements}

\item{d}{initial beta parameters ; a vector with length = # of statements}

\item{sigm}{initial sigma parameters; a matrix with # of dimensions x # of dimensions}

\item{theta}{initial theta parameters; a matrix with # of subjects x # of dimensions}

\item{B}{of iterations in each batch}

\item{J}{number of blocks}

\item{D}{number of dimensions}

\item{N}{number of participants}

\item{cor.matrix}{logical; use correlation matrix as the estimate if cor.matrix = TRUE}

\item{positive}{a logical vector indicating whether each statement is positive directional or not}

\item{fix.sigma}{logical; TRUE if sigma is not estimated}
}
\value{
A list of data.
}
\description{
kernel_4_mole
}
