% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-rank.R
\name{logLi_4_rank}
\alias{logLi_4_rank}
\title{logLi_4_rank}
\usage{
logLi_4_rank(thetai, a, d, yi, BID, prior = TRUE, sigma = NULL)
}
\arguments{
\item{thetai}{A vector; D(dimension) ability for person i.}

\item{a}{A matrix 4 x J.}

\item{d}{A matrix 4 x J.}

\item{yi}{A vector; J(block number) response for person i.}

\item{BID}{A # of statements x 3 matrix; item information,
columns are "Block", "Item" and "Dimensions".}

\item{prior}{Logical.}

\item{sigma}{Matrix D x D or NULL.}
}
\value{
a number.
}
\description{
logLi_4_rank
}
