% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MAP_data}
\alias{MAP_data}
\title{A Triple Empirical Data for Dominance Model}

\format{
A large matrix of 1391 observations containing
  information on 6 variables.
  In each block, 3 items(e.g. A/B/C) measure 3 different traits.
  A>B>C=1; A>C>B=2; B>A>C=3; B>C>A=4; C>A>B=5; C>B>A=6
}
\usage{
MAP_data
}
\description{
This data set contains a real data set
of 1391 participants on 88 triplets. In each triplet,
participants had to rank the three alternative items according
to their preference.
}
\examples{
\donttest{  
library (MIRT4FC)
Y <- data("MAP_data")
BID <- data.frame( # Block-item-demension correspondence table
 Block = rep(1:88,each=3),
 Item = rep(1:3,times=88),
 Dim = c(11, 1, 14, 15, 13, 18, 6, 2, 7, 1, 21, 24, 23, 9, 22, 5, 3, 8,
6, 18, 20, 9, 7, 5, 19, 1, 9, 14, 22, 10, 23, 21, 8, 3, 20, 10, 22,
3, 17, 4, 23, 24,
 8, 4, 24, 21, 12, 10, 9, 11, 8, 17, 12, 2, 15, 11, 14, 24, 22,
15, 13, 16, 14, 22, 4, 14, 10, 7, 6, 14, 19, 6, 4, 13, 3, 14, 15, 2,
8, 3, 11, 18, 23, 20, 24, 15,
 22, 3, 20, 1, 14, 21, 16, 18, 4, 13, 16, 18, 5, 1, 8, 23, 2,
24, 11, 19, 23, 15, 12, 11, 10, 20, 9, 21, 10, 4, 16, 7, 2, 3, 12, 16,
10, 6, 13, 16, 21, 16, 17, 20,
 10, 19, 13, 7, 5, 15, 11, 23, 24, 8, 6, 11, 2, 19, 15, 17, 20,
18, 9, 7, 12, 5, 9, 7, 22, 17, 24, 16, 6, 17, 13, 23, 7, 15, 17, 5,
8, 19, 14, 18, 3, 12, 22, 4, 5,
 21, 13, 1, 23, 9, 13, 11, 22, 9, 19, 21, 8, 21, 12, 6, 16, 1,
23, 9, 1, 19, 2, 6, 11, 18, 24, 10, 7, 18, 17, 5, 7, 3, 22, 4, 2,
8, 20, 17, 15, 8, 3, 24, 12, 10, 22,
 2, 11, 23, 19, 19, 13, 17, 6, 20, 24, 9, 17, 5, 20, 12, 6, 19,
18, 16, 15, 21, 7, 5, 1, 18, 2, 4, 14, 1, 13, 12, 16, 12, 2, 20, 4,
5, 10, 4, 1, 21, 14, 3))
######## Item parameter estimation
fit <- iStEM (Y, BID, maxitr = 150, blocksize = 3, res = 'rank', fix.sigma = TRUE)
print(fit)
}
}
\keyword{datasets}
