% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-3-pick.R
\name{P.Yi_3_pick}
\alias{P.Yi_3_pick}
\title{P.Yi_3_pick}
\usage{
P.Yi_3_pick(a, d, theta, yi)
}
\arguments{
\item{a}{matrix 3 x J}

\item{d}{matrix 3 x J}

\item{theta}{matrix 3 x J  ability for each dimension on each item}

\item{yi}{vector of J}
}
\value{
A 3*J matrix.
}
\description{
probability of yi - length of J p(yi|theta,...) ---- verified
}
