% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-rank.R
\name{P.Yi_4_rank}
\alias{P.Yi_4_rank}
\title{P.Yi_4_rank}
\usage{
P.Yi_4_rank(a, d, theta, yi)
}
\arguments{
\item{a}{matrix 4 x J}

\item{d}{matrix 4 x J}

\item{theta}{matrix 4 x J  ability for each dimension on each item}

\item{yi}{vector of J}
}
\value{
A 4*J matrix.
}
\description{
probability of yi - length of J p(yi|theta,...) ---- verified
}
