% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-3-rank.R
\name{P.Yj_3_rank}
\alias{P.Yj_3_rank}
\title{P.Yj_3_rank}
\usage{
P.Yj_3_rank(aj, dj, thetaj, yj)
}
\arguments{
\item{aj}{vector of length 3}

\item{dj}{vector of length 3}

\item{thetaj}{matrix of N x 3}

\item{yj}{vector of length N}
}
\value{
a number.
}
\description{
sum_i log P(yij|theta_i,aj,dj)
}
