% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-pick.R
\name{P.Yj_4_pick}
\alias{P.Yj_4_pick}
\title{P.Yj_4_pick}
\usage{
P.Yj_4_pick(aj, dj, thetaj, yj)
}
\arguments{
\item{aj}{vector of length 4}

\item{dj}{vector of length 4}

\item{thetaj}{matrix of N x 4}

\item{yj}{vector of length N}
}
\value{
a number.
}
\description{
sum_i log P(yij|theta_i,aj,dj)
}
