% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-rank.R
\name{P.Yj_4_rank}
\alias{P.Yj_4_rank}
\title{P.Yj_4_rank}
\usage{
P.Yj_4_rank(aj, dj, thetaj, yj)
}
\arguments{
\item{aj}{vector of length 4}

\item{dj}{vector of length 4}

\item{thetaj}{matrix of N x 4}

\item{yj}{vector of length N}
}
\value{
a number.
}
\description{
sum_i log P(yij|theta_i,aj,dj)
}
