% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-mole.R
\name{param.vec.2.ads_4_mole}
\alias{param.vec.2.ads_4_mole}
\title{param.vec.2.ads_4_mole}
\usage{
param.vec.2.ads_4_mole(pv, J, D, sigm = NULL, fix.sigma = FALSE)
}
\arguments{
\item{pv}{the vector of all parameters (a,d,sig)}

\item{J}{number of blocks}

\item{D}{number of dimension}

\item{sigm}{initial sigma parameters; a matrix with # of dimensions x # of dimensions}

\item{fix.sigma}{logical; TRUE if sigma is not estimated}
}
\value{
A list of data.
}
\description{
function used to change the format of the results
}
