% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-3-rank.R
\name{Pj_3_rank}
\alias{Pj_3_rank}
\title{Pj_3_rank}
\usage{
Pj_3_rank(j, theta, aj = NULL, dj = NULL, BID = NULL, item.par = NULL)
}
\arguments{
\item{j}{block number}

\item{theta}{matrix N x D}

\item{aj}{matrix 1 x 3}

\item{dj}{matrix 1 x 3}

\item{BID}{block,item,dimension table}

\item{item.par}{a data.frame including a,d parameter}
}
\value{
a 3*N matrix.
}
\description{
Calculate probabilities of choosing each option out of
             6 alternatives of block j by all N persons: 6 x N matrix
}
