% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-4-pick.R
\name{Pj_4_pick}
\alias{Pj_4_pick}
\title{Pj_4_pick}
\usage{
Pj_4_pick(j, theta, aj = NULL, dj = NULL, BID = NULL, item.par = NULL)
}
\arguments{
\item{j}{block number}

\item{theta}{matrix N x D}

\item{aj}{matrix 1 x 4}

\item{dj}{matrix 1 x 4}

\item{BID}{block,item,dimension table}

\item{item.par}{a data.frame including a,d parameter}
}
\value{
4*N matrix.
}
\description{
Calculate probabilities of choosing each option out of
             4 alternatives of block j by all N persons: 4 x N matrix
}
