% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stEM-4-pick.R
\name{stEM_4_pick}
\alias{stEM_4_pick}
\title{stEM_4_pick}
\usage{
stEM_4_pick(
  Y,
  BID,
  positive = rep(TRUE, nrow(BID)),
  M = 10,
  B = 20,
  a = NULL,
  d = NULL,
  item.par = NULL,
  sigma = NULL,
  theta = NULL,
  fix.sigma = FALSE,
  burnin.maxitr = 40,
  maxitr = 500,
  eps1 = 1.5,
  eps2 = 0.4,
  frac1 = 0.2,
  frac2 = 0.5,
  cores = NULL
)
}
\arguments{
\item{Y}{A # of subjects x # of blocks matrix; item responses}

\item{BID}{A # of statements x 3 matrix; item information,
columns are "Block", "Item" and "Dimensions"}

\item{positive}{A logical vector; indicating whether each statement is
positive directional or not}

\item{M}{A number; # of batch}

\item{B}{A number; # of iterations in each batch}

\item{a}{A vector; length = # of statements, initial alpha parameters}

\item{d}{A vector; length = # of statements, initial beta parameters}

\item{item.par}{A data frame; initial parameters for a and d}

\item{sigma}{A # of dimensions x # of dimensions matrix; initial sigma
parameters}

\item{theta}{A # of subjects x # of dimensions matrix; initial theta parameters}

\item{fix.sigma}{Logical; TRUE if sigma is estimated}

\item{burnin.maxitr}{A number; max burn-in allowed}

\item{maxitr}{A number; max iterations allowed}

\item{eps1}{A number; stability criteria}

\item{eps2}{A number; convergence criterion}

\item{frac1}{A number; cutoffs for calculating Geweke z}

\item{frac2}{A number; cutoffs for calculating Geweke z}

\item{cores}{A number; number of cores}
}
\value{
A list of data; including estimated a, d, and sigm parameters,
   total batch number, final chain size, burn-in size, time
}
\description{
stEM_4_pick
}
