% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.est.R
\name{theta.est}
\alias{theta.est}
\title{Provide the function of estimating latent traits.}
\usage{
theta.est(
  Y,
  a = NULL,
  d = NULL,
  t = NULL,
  BID,
  sigma = NULL,
  prior = TRUE,
  blocksize = 3,
  res = "rank",
  model = "2PL"
)
}
\arguments{
\item{Y}{A # of subjects * # of blocks matrix; item responses.}

\item{a}{A vector; length = # of statements, initial alpha parameters.}

\item{d}{A vector; length = # of statements, initial beta parameters.}

\item{BID}{A # of statements * 3 matrix; item information,
columns are "Block", "Item" and "Dimensions".}

\item{sigma}{A # of dimensions * # of dimensions matrix; initial sigma
parameters.}

\item{blocksize}{A number; block size of FC(2/3/4).}

\item{res}{A string; response format('pick'/'rank'/'mole'),
pick-2(blocksize=2)/rank-2/mole-2 are equivalent, rank-3/mole-3 are equivalent.}

\item{model}{A string; FC model('2PL').}
}
\description{
Estimating latent traits for multivariate FC item response theory model.
}
\value{
A # of subjects x # of block number matrix
}
\examples{
\donttest{
library (MIRT4FC)
D <- 6
nitem.per.dim <- 10
nblock <- D * nitem.per.dim / 3
item.par <- data.frame(a=seq_len(D*nitem.per.dim))
item.par <- within(item.par,{
 a <- runif(D*nitem.per.dim,0.7,3)
 b <- rnorm(D*nitem.per.dim)
 d <- a*b
})
a=matrix(item.par$a,nrow=3,ncol = 20)
d=matrix(item.par$d,nrow=3,ncol = 20)
BID <- data.frame(Block=rep(1:nblock,each=3),
 Item=rep(1:3,nblock),
 Dim=c(combn(D,3)[,sample(choose(D,3),nblock,replace = TRUE)]))
item.par$d <- c(t(aggregate(item.par$d,by=list(BID$Block),function(x)x-mean(x))[,-1]))
N <- 1000
v <- matrix(0,D,D)
diag(v) <- 1
eigen(v)$values
theta <- mvnfast::rmvn(N,rep(0,each = D),sigma = v)
Y <- data.sim(item.par,theta,BID,blocksize = 3,res = "rank")
theta_est <- theta.est(Y,a,d,BID=BID,sigma=v,prior=TRUE,blocksize=3,res='rank',model='2PL')
}
}
