% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sim.R
\name{data.sim}
\alias{data.sim}
\title{Data.sim}
\usage{
data.sim(
  item.par = item.par,
  theta = theta,
  BID = BID,
  blocksize = 3,
  res = "rank",
  model = "2PL"
)
}
\arguments{
\item{item.par}{A data frame; parameters for a and d.}

\item{theta}{A # of subjects * # of dimensions matrix; theta parameters.}

\item{BID}{A # of statements * 3 matrix; item information,
columns are "Block", "Item" and "Dimensions".}

\item{blocksize}{A number; block size of FC(2/3/4).}

\item{res}{A string; response format('pick'/'rank'/'mole'),
pick-2(blocksize=2)/rank-2/mole-2 are equivalent, rank-3/mole-3 are equivalent.}
}
\value{
A # of subjects * # of block number matrix.
}
\description{
Data.sim
}
\examples{
D <- 6
nitem.per.dim <- 10
nblock <- D * nitem.per.dim / 3
set.seed(123456)
item.par <- data.frame(a=seq_len(D*nitem.per.dim))
item.par <- within(item.par,{
  a <- runif(D*nitem.per.dim,0.7,3)
  b <- rnorm(D*nitem.per.dim)
  d <- a*b
})
BID <- data.frame(Block=rep(1:nblock,each=3),
                  Item=rep(1:3,nblock),
                  Dim=c(combn(D,3)[,sample(choose(D,3),nblock,replace = TRUE)]))
item.par$d <- c(t(aggregate(item.par$d,by=list(BID$Block),function(x)x-mean(x))[,-1]))
N <- 1000
v <- matrix(0.5,D,D)
diag(v) <- 1
theta <- mvnfast::rmvn(N,seq(-1,1,length.out = D),sigma = v)
Y <- data.sim(item.par,theta,BID,blocksize=3,res='rank')

}
