% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MAP_data}
\alias{MAP_data}
\title{MAP_data}
\format{
A data frame of 200 observations containing
  information on 6 variables.
  In each block, 3 items(e.g. A/B/C) measure 3 different traits.
  A>B>C=1; A>C>B=2; B>A>C=3; B>C>A=4;C>A>B=5; C>B>A=6
}
\usage{
MAP_data
}
\description{
This data set contains a real data set
of 1391 participants on 88 triplets. In each triplet,
participants had to rank the three alternative items according
to their preference.
}
\keyword{datasets}
