% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.est.R
\name{theta.est}
\alias{theta.est}
\title{theta.est}
\usage{
theta.est(
  Y,
  a = NULL,
  d = NULL,
  t = NULL,
  BID,
  sigma = NULL,
  prior = TRUE,
  blocksize = 3,
  res = "rank",
  model = "2PL"
)
}
\arguments{
\item{Y}{A # of subjects * # of blocks matrix; item responses.}

\item{a}{A vector; length = # of statements, initial alpha parameters.}

\item{d}{A vector; length = # of statements, initial beta parameters.}

\item{BID}{A # of statements * 3 matrix; item information,
columns are "Block", "Item" and "Dimensions".}

\item{sigma}{A # of dimensions * # of dimensions matrix; initial sigma
parameters.}

\item{blocksize}{A number; block size of FC(2/3/4).}

\item{res}{A string; response format('pick'/'rank'/'mole'),
pick-2(blocksize=2)/rank-2/mole-2 are equivalent, rank-3/mole-3 are equivalent.}

\item{positive}{A logical vector; indicating whether each statement is
positive directional or not.}

\item{M}{A number; # of batch.}

\item{B}{A number; # of iterations in each batch.}

\item{item.par}{A data frame; initial parameters for a and d.}

\item{theta}{A # of subjects * # of dimensions matrix; initial theta parameters.}

\item{fix.sigma}{Logical; TRUE if sigma is estimated.}

\item{burnin.maxitr}{A number; max burn-in allowed.}

\item{maxitr}{A number; max iterations allowed.}

\item{eps1}{A number; stability criteria.}

\item{eps2}{A number; convergence criterion.}

\item{frac1}{A number; cutoffs for calculating Geweke z.}

\item{frac2}{A number; cutoffs for calculating Geweke z.}

\item{cores}{A number; number of cores.}
}
\value{
A list of data; including estimated a, d, and sigm parameters,
   total batch number, final chain size, burn-in size, time.
}
\description{
theta.est
}
\examples{
\donttest{
D <- 6
nitem.per.dim <- 10
nblock <- D * nitem.per.dim / 3
set.seed(123456)
item.par <- data.frame(a=seq_len(D*nitem.per.dim))
item.par <- within(item.par,{
  a <- runif(D*nitem.per.dim,0.7,3)
  b <- rnorm(D*nitem.per.dim)
  d <- a*b
})
BID <- data.frame(Block=rep(1:nblock,each=3),
                  Item=rep(1:3,nblock),
                  Dim=c(combn(D,3)[,sample(choose(D,3),nblock,replace = TRUE)]))
item.par$d <- c(t(aggregate(item.par$d,by=list(BID$Block),function(x)x-mean(x))[,-1]))
N <- 1000
v <- matrix(0.5,D,D)
diag(v) <- 1
theta <- mvnfast::rmvn(N,seq(-1,1,length.out = D),sigma = v)
Y <- data.sim(item.par,theta,BID,blocksize=3,res='rank')
x <- StEM(Y,BID,maxitr = 100,blocksize=3,res='rank',fix.sigma = TRUE,cores=1)
}
}
